Add files via upload
Two scripts:
sd-crop-1k.py crops larger images automatically to 1024x1024px pieces to include details of larger images in the training set. It's designed to be used together with the other script, sd-resize-1m.py which resizes all images to 1M pixels for SD training, and checks that the dimensions are divisible by 64 pixels. If not, contrary to what the comments say, it just resizes the image for now, so aspect ratio will change.